QUALITY: Refactor entry into program
