+++ src/commands/run.js
@@ -1,4 +1,4 @@
-const runCommand = require('../utilities').runCommand;
+const runCommand = require('../utils').runCommand;
 
 const execute = function(args, callback) {
   runCommand('./program')
