+++ src/gui/index.js
@@ -1,0 +1,7 @@
+const execute = function () {
+  console.error("GUI not yet implemented!")
+}
+
+module.exports = {
+  execute: execute
+}
